refactor!: move TigerBeetlex to TigerBeetlex.Client
Incidentally, rename TigerBeetlex.connect/2 to TigerBeetlex.Client.new/2

Use a deprecation warning to avoid breaking backwards compatibility
immediately. The TigerBeetlex module will be removed before the first
release.
